Fix Bug in assoc_path
Consider the case when we are using `inverse_topology` to get the
absolute form of an update with two ports, `global` and `port`. Say the
topology maps `global` to `boundary` and `port` to
`('boundary', 'store')`. Before this commit, if inverse_topology handled
`port` and then `global`, it would first create a dictionary with the
value of `port` at `('boundary', 'global', 'store')`. However, next when
handling `global`, it would overwrite `store` with the value of `global`
instead of merging the dictionaries.

This commit fixes `assoc_path` and creates a test to check for any
regressions.