Avoid Calculating SEM of Single-Element Arrays
Using scipy.stats.sem to calculate the SEM of single-element arrays
causes RuntimeWarnings:
        * RuntimeWarning: Degrees of freedom <= 0 for slice
        * RuntimeWarning: invalid value encountered in double_scalars
This commit skips scipy and sets the SEM of single-element arrays to
zero for computing error bands in the colony metrics plots. This is
reasonable because for single values, the spread of the distribution,
which is what the error bands represent, is zero.