Fix Analysis Script to Match Tags Plot
Now that the tags plot expects a full path to the tagged molecule, we
need to change the analysis script accordingly. Here we do so. This also
requires that the CSV of tagged molecules be updated to include as the
first item of every row the store (`boundary`) to access.