Handle Updater Dictionaries in Experiment & Store
Now, updates can come in any of the following forms:

* Single value:

  ```
  update = {'port': value}
  ```

* Specific Updater

  ```
  update = {'_updater': updater_func, '_value': value}
  ```

* Specific Updater (updater string)

  ```
  update = {'_updater': 'my_updater', '_value': value}
  ```

* Updater with Port Mapping

  ```
  update = {
      '_value': value,
      '_updater': {
          'updater': updater_func,
          'port_mapping': {
              'updaterPort1': 'processPort1',
              'updaterPort2': 'processPort2',
          }
      }
  }

* Updater with Port Mapping (updater string)

  ```
  update = {
      '_value': value,
      '_updater': {
          'updater': 'my_updater',
          'port_mapping': {
              'updaterPort1': 'processPort1',
              'updaterPort2': 'processPort2',
          }
      }
  }

This commit changes the Experiment and Store classes to handle all of
these. Now we store updates as tuples of the update object (calling
.get() gives the update dictionary), the process topology, and the state
Store used to get the process's update's states. We use these when
applying the update to get the state specified by `port_mapping`. We
also update the code that handles defaults to work with numpy array
defaults.